Changes to support compatibility pack v4.
This compatibility pack allows use of honeycomb features on older SDKs.
App now runs on 2.3.3 emulator as well as 3.0 emulator.
Set min SDK level to 10 (2.3.3) for now, and got it to run on the galaxyS2.
We can lower the level to get it to run on other phones as well later.
But for now, I am able to test on emulator and phone, so will leave it as is.